Rework BSFT file is updated in BAA file.
Previously, the embedded `.bsft` file in `GCKart.baa` was updated by
appending another copy of it, and overriding its data offset in the
header entry. This was not ideal, but it was functional.

Now, `GCKart.baa` is unpacked first, then its `.bsft` is rewritten, and
the BAA file is re-packed.